+++ src/Parser/Evaluated/NaturalParser.php
@@ -15,6 +15,7 @@ use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\ImageGroup;
 use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\InTheNews;
 use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\Map;
 use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\Classical\ClassicalCardsResult;
+use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\MapLegacy;
 use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\PeopleAlsoAsk;
 use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\SearchResultGroup;
 use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\TopStoriesCarousel;
@@ -45,6 +46,7 @@ class NaturalParser extends AbstractParser
             new ClassicalWithLargeVideo(),
             new InTheNews(),
             new Map(),
+            new MapLegacy(),
             new AnswerBox(),
             new Flight(),
             new PeopleAlsoAsk()