+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 ## 0.2.3
 
-*2017-12-10*
+*2017-12-xx*
 
 * Features:
     * Added parsing for people also ask results